LPD-95940 Extract the assistant categorization flow into a hook
