+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/content_editor/components/ContentEditorSidePanel.tsx
@@ -12,34 +12,21 @@ import {useIsMounted} from '@liferay/frontend-js-react-web';
 import {datetimeUtils} from '@liferay/object-js-components-web';
 import {LiferayEditorConfig} from 'frontend-editor-ckeditor-web';
 import {openToast} from 'frontend-js-components-web';
-import {fetch, objectToFormData, sub} from 'frontend-js-web';
+import {fetch, objectToFormData} from 'frontend-js-web';
 import React, {useCallback, useEffect, useRef, useState} from 'react';
 
-import CategorizationSuggestionService from '../../common/services/CategorizationSuggestionService';
 import VocabularyService from '../../common/services/VocabularyService';
 import {IAssetObjectEntry} from '../../common/types/AssetType';
 import focusInvalidElement from '../../common/utils/focusInvalidElement';
-import {
-	AUTO_CATEGORIZE_AGENT,
-	CATEGORIZE_EVENT,
-	COMMIT_EVENT,
-	CategorizationAction,
-	CategorizationCommitPayload,
-	CategorizeEventPayload,
-	GENERATE_TAGS_AGENT,
-	OPEN_CATEGORIZATION_PANEL_EVENT,
-	REQUEST_CATEGORIZE_EVENT,
-	RequestCategorizePayload,
-} from '../../main_view/info_panel/components/categorizationAgentEvents';
 import ObjectEntryService from '../../main_view/info_panel/services/ObjectEntryService';
 import {Comment} from '../services/CommentService';
-import getEditedContent from '../utils/getEditedContent';
 import {EVENT_VALIDATE_FORM} from './ContentEditorToolbar';
 import {dateConfig, toMomentDate, toServerISOFormat} from './ScheduleField';
 import CategorizationPanel from './panels/CategorizationPanel';
 import CommentsPanel from './panels/CommentsPanel';
 import GeneralPanel from './panels/GeneralPanel';
 import SchedulePanel from './panels/SchedulePanel';
+import useAssistantCategorization from './useAssistantCategorization';
 
 type Props = {
 	addCommentURL: string;
@@ -331,192 +318,26 @@ function SidePanel(props: SidePanelProps) {
 	const [hasError, setHasError] = useState<boolean>(false);
 	const [panel, setPanel] = useState<React.Key | null>(null);
 
-	const categorizationFieldsRef = useRef(props.categorizationFields);
-	const panelRef = useRef(panel);
-
-	useEffect(() => {
-		categorizationFieldsRef.current = props.categorizationFields;
-	}, [props.categorizationFields]);
-
-	useEffect(() => {
-		panelRef.current = panel;
-	}, [panel]);
+	const openCategorizationPanel = useCallback(() => {
+		setPanel('categorization');
+	}, []);
 
 	const showErrorInPanel = useCallback((panelId: React.Key) => {
 		setPanel(panelId);
 		setHasError(true);
 	}, []);
 
-	const {assetLibraryId, cmsGroupId, contentAPIURL, onUpdateCategorization} =
-		props;
+	const {onUpdateCategorization} = props;
 
-	useEffect(() => {
-		const fireCategorize = async (actions: CategorizationAction[]) => {
-			const {data, error} =
-				await ObjectEntryService.getObjectEntry(contentAPIURL);
-
-			if (!data) {
-				if (error) {
-					console.error(error);
-				}
-
-				return;
-			}
-
-			const editedContent = await getEditedContent(
-				data.systemProperties?.objectDefinitionBrief
-					?.externalReferenceCode
-			);
-
-			actions.forEach((action) => {
-				const agent =
-					action.agent === 'categorize'
-						? AUTO_CATEGORIZE_AGENT
-						: GENERATE_TAGS_AGENT;
-
-				const payload: CategorizeEventPayload = {
-					agent,
-					cmsGroupId,
-					content: editedContent || data.contentRawText || '',
-					count: action.count,
-					scopeId:
-						agent === AUTO_CATEGORIZE_AGENT
-							? data.scopeId
-							: data.scopeId || assetLibraryId || cmsGroupId,
-					suppressUserMessage: true,
-					targets: action.targets,
-				};
-
-				if (agent === AUTO_CATEGORIZE_AGENT) {
-					payload.classNameId =
-						data.systemProperties?.objectDefinitionBrief
-							?.classNameId ?? -1;
-					payload.currentCategoryIds = (
-						data.taxonomyCategoryBriefs || []
-					).map((brief) => brief.taxonomyCategoryId);
-				}
-				else {
-					payload.currentTagNames = data.keywords || [];
-				}
-
-				Liferay.fire(CATEGORIZE_EVENT, payload);
-			});
-		};
-
-		const handleRequestCategorize = (payload: RequestCategorizePayload) => {
-			fireCategorize(payload.actions);
-		};
-
-		const handleCommit = async ({
-			agent,
-			notifyAssistantPanelOpen,
-			scopeId,
-			suggestions,
-		}: CategorizationCommitPayload) => {
-			const fields = categorizationFieldsRef.current;
-
-			if (panelRef.current === 'categorization' || !fields) {
-				return;
-			}
-
-			notifyAssistantPanelOpen?.(false);
-
-			if (agent === AUTO_CATEGORIZE_AGENT) {
-				const currentBriefs = fields.assetCategoryIds.value;
-
-				const briefs =
-					await CategorizationSuggestionService.resolveNewCategoryBriefs(
-						suggestions,
-						currentBriefs.map(
-							({taxonomyCategoryId}) => taxonomyCategoryId
-						)
-					);
-
-				if (!briefs.length) {
-					return;
-				}
-
-				const value = [...currentBriefs, ...briefs];
-
-				onUpdateCategorization([
-					'assetCategoryIds',
-					{
-						serverValue: value
-							.map(({taxonomyCategoryId}) => taxonomyCategoryId)
-							.join(','),
-						value,
-					},
-				]);
-
-				openToast({
-					message: sub(
-						Liferay.Language.get(
-							'x-categories-have-been-successfully-added-to-the-selected-content'
-						),
-						`${briefs.length}`
-					),
-					type: 'success',
-				});
-			}
-			else if (agent === GENERATE_TAGS_AGENT) {
-				const names =
-					await CategorizationSuggestionService.createTagNames(
-						suggestions,
-						{
-							assetLibraryId:
-								scopeId || assetLibraryId || cmsGroupId,
-							cmsGroupId,
-						}
-					);
-
-				const currentNames = fields.assetTagNames.value;
-
-				const newNames = [
-					...new Set(
-						names.filter((name) => !currentNames.includes(name))
-					),
-				];
-
-				if (!newNames.length) {
-					return;
-				}
-
-				const keywords = [...currentNames, ...newNames];
-
-				onUpdateCategorization([
-					'assetTagNames',
-					{serverValue: keywords.join(','), value: keywords},
-				]);
-
-				openToast({
-					message: sub(
-						Liferay.Language.get(
-							'x-tags-have-been-successfully-added-to-the-selected-content'
-						),
-						`${newNames.length}`
-					),
-					type: 'success',
-				});
-			}
-		};
-
-		const openCategorizationPanel = () => {
-			setPanel('categorization');
-		};
-
-		Liferay.on(COMMIT_EVENT, handleCommit);
-		Liferay.on(REQUEST_CATEGORIZE_EVENT, handleRequestCategorize);
-		Liferay.on(OPEN_CATEGORIZATION_PANEL_EVENT, openCategorizationPanel);
-
-		return () => {
-			Liferay.detach(COMMIT_EVENT, handleCommit);
-			Liferay.detach(REQUEST_CATEGORIZE_EVENT, handleRequestCategorize);
-			Liferay.detach(
-				OPEN_CATEGORIZATION_PANEL_EVENT,
-				openCategorizationPanel
-			);
-		};
-	}, [assetLibraryId, cmsGroupId, contentAPIURL, onUpdateCategorization]);
+	useAssistantCategorization({
+		assetLibraryId: props.assetLibraryId,
+		categorizationFields: props.categorizationFields,
+		cmsGroupId: props.cmsGroupId,
+		contentAPIURL: props.contentAPIURL,
+		onOpenCategorizationPanel: openCategorizationPanel,
+		onUpdateCategorization,
+		panel,
+	});
 
 	useEffect(() => {
 		const validateScheduleFields = ({event}: {event: MouseEvent}) => {
